fix(abi): correct x86_64 Stat struct layout and waitpid encoding
TEAM_446: Fix two critical Linux ABI compliance bugs:

1. x86_64 Stat struct layout (CRITICAL):
   - Created architecture-specific Stat structs
   - aarch64: 128 bytes (asm-generic layout)
   - x86_64: 144 bytes with st_nlink (u64) before st_mode
   - Added compile-time size assertions
   - Updated API: new_pipe/from_inode_data take i64 for blksize

2. waitpid status encoding (MEDIUM):
   - Changed from raw exit_code to (exit_code & 0xFF) << 8
   - WIFEXITED() and WEXITSTATUS() macros now work correctly
   - Shell $? variable reports correct exit codes

Additional fixes:
- Added #[cfg(target_arch = "x86_64")] for Open/Socketpair syscalls
- Made write_to_user_buf() public for levitate crate access

+++ syscall/src/fs/stat.rs
@@ -24,8 +24,9 @@ pub fn sys_fstat(fd: usize, stat_buf: usize) -> SyscallResult {
             Err(e) => return Err(e.to_errno()),
         },
         // TEAM_258: Use constructor for architecture independence
+        // TEAM_446: Changed to i64 for x86_64 ABI compatibility
         FdType::PipeRead(_) | FdType::PipeWrite(_) => {
-            Stat::new_pipe(los_vfs::pipe::PIPE_BUF_SIZE as i32)
+            Stat::new_pipe(los_vfs::pipe::PIPE_BUF_SIZE as i64)
         }
         // TEAM_258: Use constructor for architecture independence
         FdType::PtyMaster(_) | FdType::PtySlave(_) => {
